Rename: 수강권 컴포넌트, 페이지 tickets로 폴더명 변경
